services/hunter/requirements.txt
alpaca-py
kafka-python
python-dotenv
prometheus-client
